Changes how items are added for a subscription
This changes how subscriptions items are added. These were previously
added as invoice items during one of the renewal webhook events. But
this removes that functionality and instead adds items directly to a
subscription.

Note that going with this approach means we need to add a generic
product in Stripe for each of our WC order item types.